requirements/base.txt
django-authtools==1.6.0
django-braces==1.13.0
django-crispy-forms==1.7.2
django-environ==0.4.5
django==2.2.24
easy-thumbnails==2.5
pillow==8.2.0
pytz==2019.3
django_select2==7.2.0
psycopg2==2.8.4
django-leaflet==0.26.0
geopy==1.21.0
django-countries==5.5
django_summernote
django-rated-reviews
django-cookie-law
djangorestframework==3.11.2
djoser==2.0.3
django-rest-swagger==2.2.0
django-oidc-provider==0.7.0
drf-yasg==1.17.1
django-oidc-provider==0.7.0
django-recaptcha==2.0.6
django-machina==1.1.3
django-active-link
django-cron
django-crontab
django-ckeditor==6.0.0
django-oauth-toolkit==1.3.3
whoosh
argon2_cffi
